==> Abseneeism_business_case/01_Absenteeism_business_case.ipynb <==
# Absenteeism business case for better HR management

# Import library to load data.

import pandas as pd

# Load data.

data_raw = pd.read_csv('Absenteeism-data.csv')

# Check some records.
# if you want to see all rows and columns, use the following code:
    # pd.options.display.max_columns = None
    # pd.options.display.max_rows = None
    # display(data_name)

data_raw.head()

# Copy data for manipulation.

data_copy = data_raw.copy()

# Check if copy worked.

data_copy.head()

# Check general info for data.

data_copy.info()

# Basic descriptive statistics from data.

data_copy.describe(include = 'all')

# Drop 'ID' column since it will not be used for analysis.

data_copy = data_copy.drop(['ID'], axis = 1)

# Check if drop worked.

data_copy.head()

# How many distinct reasons for absence are there?

sorted(pd.unique(data_copy['Reason for Absence']))

# There is not a '20' reason, look deeper into that.

# How many distinct reasons for absence are there?

len(pd.unique(data_copy['Reason for Absence']))

# Create dummy data for reason of absence.

reason_dummies = pd.get_dummies(data_copy['Reason for Absence'])

# Check if data was created correctly.

reason_dummies.head()

# Check for nulls and outliers in reason for absence.

# Add new column for check.

reason_dummies['check'] = reason_dummies.sum(axis = 1)

# Check new column.

reason_dummies

# Check for missing values.

reason_dummies['check'].sum(axis = 0)

# No missing values.

# Check for outliers.

pd.unique(reason_dummies['check'])

# No outliers, since employess can only be absent for one reason.

# Drop check column since it is no longer useful.

reason_dummies = reason_dummies.drop(['check'], axis = 1)

# Check if drop worked correctly.

reason_dummies.head()

# Drop reason 0 to avoid multicolinearity.

reason_dummies = pd.get_dummies(data_copy['Reason for Absence'], drop_first = True)

# Check new data.

reason_dummies.head()

# Group reasons for absence since 28 categories is too much.

# Drop column 'Reason for absence' since it will be replaced by dummies.

data_copy = data_copy.drop(['Reason for Absence'], axis = 1)

# Check if drop worked correctly.

data_copy.head()

# Group dummies into 4 groups.

absence_reason_1 = reason_dummies.loc[:, 1:14].max(axis = 1)
absence_reason_2 = reason_dummies.loc[:, 15:17].max(axis = 1)
absence_reason_3 = reason_dummies.loc[:, 18:21].max(axis = 1)
absence_reason_4 = reason_dummies.loc[:, 22:].max(axis = 1)

# Check columns to see if grouping worked.

absence_reason_1.head()

absence_reason_2.head()

absence_reason_3.head()

absence_reason_4.head()

# Attach grouped reasons for absence columns to full data.

data_copy = pd.concat([data_copy, 
                      absence_reason_1, 
                      absence_reason_2, 
                      absence_reason_3,
                      absence_reason_4],
                     axis = 1)

# Check if columns were properly added.

data_copy.head()

# Change names of columns for reasons for absence.

data_copy.columns.values

column_names = ['Date', 'Transportation Expense', 'Distance to Work', 'Age',
       'Daily Work Load Average', 'Body Mass Index', 'Education',
       'Children', 'Pets', 'Absenteeism Time in Hours', 'absence_reason_1', 
                'absence_reason_2', 'absence_reason_3', 'absence_reason_4']

data_copy.columns = column_names

# Check if name change worked.

data_copy.head()

# Reorder columns.

column_names_reorder = ['absence_reason_1', 'absence_reason_2', 'absence_reason_3', 
                        'absence_reason_4', 'Date', 'Transportation Expense', 
                        'Distance to Work', 'Age','Daily Work Load Average', 
                        'Body Mass Index', 'Education', 'Children', 'Pets', 
                        'Absenteeism Time in Hours']

data_copy = data_copy[column_names_reorder]

data_copy.head()

# Create a checkpoint

data_reason_mod = data_copy.copy()

data_reason_mod.head()

# Take a deep dive for 'Date'.

type(data_reason_mod['Date'][0])

# Variable is a string. Change to date format.

data_reason_mod['Date'] = pd.to_datetime(data_reason_mod['Date'],
                                        format = '%d/%m/%Y')

# Check if change is correct.

data_reason_mod.info()

type(data_reason_mod['Date'][0])

# Extract month value from date.

# Check if month is display correctly on its own.

data_reason_mod['Date'][0].month

list_months = []

for i in range(data_reason_mod.shape[0]):
    list_months.append(data_reason_mod['Date'][i].month)

# Check if columns was created correctly.

list_months[0:5]

len(list_months)

# Add new month column created to data.

data_reason_mod['Month value'] = list_months

# Check if column was correctly added.

data_reason_mod.head()

# Extract the day of the week.

# Check if weekday is displayed correctly.

data_reason_mod['Date'][699].weekday()

def date_to_weekday(date_value):
    return date_value.weekday()

data_reason_mod['Day of the week'] = data_reason_mod['Date'].apply(date_to_weekday)

# Check if new weekday columns was created correctly.

data_reason_mod.head()

# Drop 'Date' column since it will be not be used further.

data_reason_mod = data_reason_mod.drop(['Date'], axis = 1)

# Check if drop worked.

data_reason_mod.head()

# Reorder columns so month and weekday are where date used to be.

data_reason_mod.columns

reason_date_columns = ['absence_reason_1', 'absence_reason_2', 'absence_reason_3',
       'absence_reason_4', 'Month value', 'Day of the week', 'Transportation Expense', 
    'Distance to Work', 'Age', 'Daily Work Load Average', 'Body Mass Index', 'Education', 
    'Children', 'Pets', 'Absenteeism Time in Hours']

data_reason_mod = data_reason_mod[reason_date_columns]

# Check if reorder worked.

data_reason_mod.head()

# Create checkpoint for new data.

data_reason_date_mod = data_reason_mod.copy()

data_reason_date_mod.head()

# Check type for transportation expense (dollars per month), distance to work (km),
# age (years),daily workload average (minutes) and body mass index.

data_reason_date_mod.info()

# Check values for education.

data_reason_date_mod['Education'].unique()

# 1 stands for high school as max education.
# 2 = graduate.
# 3 = postgraduate.
# 4 = masters or doctor.

# Check distribution.

data_reason_date_mod['Education'].value_counts()

# Group education above high school into single group and change mapping of variable.

data_reason_date_mod['Education'] = data_reason_date_mod['Education'].map({1:0,
                                                                          2:1,
                                                                          3:1,
                                                                          4:1})

# Chek if new mapping worked correctly.

data_reason_date_mod['Education'].value_counts()

# Create checkpoint for new data.

data_preprocess = data_reason_date_mod.copy()

# Check if copy is correct.

data_preprocess.head()

# Save preprocessed data to csv file.

data_preprocess.to_csv('Absenteeism_preprocess.csv', index = False)
