==> Abseneeism_business_case/02_Absenteeism_business_case_model.ipynb <==
# Statistical modelling for absenteeism

# Import libraries.

import pandas as pd
import numpy as np

# Load data we preprocessed earlier.

data_0 = pd.read_csv('Absenteeism_preprocess.csv')

# Check if data loaded correctly.

data_0.head()

data_0.info()

data_0.describe(include = 'all')

# Create targets for logistic regression.

# People above median of absenteeism are considered abnormal.
# By using median we balance the data so the model can be more properly trained.

# Check median absenteeism hours.

data_0['Absenteeism Time in Hours'].median()

targets = np.where(data_0['Absenteeism Time in Hours'] > 
                   data_0['Absenteeism Time in Hours'].median(), 1, 0)

targets[0:10]

# Add new column to data.

data_0['Excessive Absenteeism'] = targets

# Check if column appended correctly.

data_0.head()

targets.sum() / targets.shape[0]

# Drop variables that will not be used further.

data_targets = data_0.drop(['Absenteeism Time in Hours',
                           'Day of the week',
                           'Daily Work Load Average', 
                           'Distance to Work'], 
                           axis = 1)

# Check new data.

data_targets.head()

# Select inputs for regression.

data_targets.shape

inputs = data_targets.iloc[:, :-1]

# Check if inputs extracted correctly.

inputs.head()

# Standardize data.

# Import modules for standardization.

from sklearn.preprocessing import StandardScaler
from sklearn.base import BaseEstimator, TransformerMixin

# create the Custom Scaler class

class CustomScaler(BaseEstimator,TransformerMixin): 
    
    def __init__(self,columns):
        self.scaler = StandardScaler()
        self.columns = columns
        self.mean_ = None
        self.var_ = None
        
    def fit(self, X, y=None):
        self.scaler.fit(X[self.columns], y)
        self.mean_ = np.mean(X[self.columns])
        self.var_ = np.var(X[self.columns])
        return self
    
    def transform(self, X, y=None, copy=None):
        init_col_order = X.columns
        X_scaled = pd.DataFrame(self.scaler.transform(X[self.columns]), columns=self.columns)
        X_not_scaled = X.loc[:,~X.columns.isin(self.columns)]
        return pd.concat([X_not_scaled, X_scaled], axis=1)[init_col_order]

inputs_to_omit = ['absence_reason_1', 'absence_reason_2',
                      'absence_reason_3', 'absence_reason_4','Education']

inputs_to_scale = [x for x in inputs.columns.values if x not in inputs_to_omit]

absenteeism_scaler = CustomScaler(inputs_to_scale)

absenteeism_scaler.fit(inputs)

inputs_scaled = absenteeism_scaler.transform(inputs)

pd.DataFrame(inputs_scaled)

inputs_scaled.shape

# Split data into train and test samples.

# Import module for splitting.

from sklearn.model_selection import train_test_split

# Split the data.

x_train, x_test, y_train, y_test = train_test_split(inputs_scaled, 
                                                    targets, 
                                                    train_size = 0.8,
                                                    shuffle = True,
                                                    random_state = 20)

# Import modules for logistic regression.

from sklearn.linear_model import LogisticRegression
from sklearn import metrics

# Train the model.

reg = LogisticRegression()

reg.fit(x_train, y_train)

# Check accuracy of model.

reg.score(x_train, y_train)

model_output = reg.predict(x_train)

np.sum((model_output == y_train)) / model_output.shape[0]

# Get intercept and coeffcients.

reg.intercept_

feature_name = inputs.columns.values

summary_table = pd.DataFrame(columns = ['feature_name'],
                            data = feature_name)

summary_table['coefficient'] = np.transpose(reg.coef_)

summary_table

summary_table.index = summary_table.index + 1

summary_table.loc[0] = ['Intercept', reg.intercept_[0]]

summary_table = summary_table.sort_index()

summary_table

# Interpreting the coefficients.

summary_table['odds_ratio'] = np.exp(summary_table.coefficient)

summary_table

summary_table.sort_values('odds_ratio', ascending = False)

# A quick summary of the meaning of reasons.
# Reason 0, the base, means no justification for absence was given.
# Reason 1 comprises various diseases.
# Reason 2 is related to pregnancy and giving birth.
# Reason 3 is some kind of poisoning.
# Reason 4 covers light diseases.

# A way to interpret the coefficients for the reasons is:
# Someone with some kind of poisoning is 22 times more likely to be absent from work
# from someone who didn´t specify a reason.

# Testing the model.

reg.score(x_test, y_test)

# Test accuracy is around 75%, lower than train at 77% but still good.

# Get predicted probability.

pred_prob = reg.predict_proba(x_test)

pred_prob[0:10]

# First columns is probability of being 0, second columns is probability of being 1.

# Getting the probabilites of falling in 1.

pred_prob[0:10, 1]

# Save the model for later use.

# Import module for saving.

import pickle

# Save model as pickle object.

with open('model', 'wb') as file:
    pickle.dump(reg, file)

# Save scaler.

with open('scaler', 'wb') as file:
    pickle.dump(absenteeism_scaler, file)
