==> Abseneeism_business_case/04_Absenteeism_module_predictions.ipynb <==
# Load module prepared with all necessary steps
from Absenteeism_module import *

# Create model from module class and call methods to be used
model = absenteeism_model('model', 'scaler')

# Check data for which methods will be uses
data = pd.read_csv('Absenteeism_new_data.csv')
data.head()

# Data is in correct format to apply methods directly, procede

# Use module to load, pre process and apply model to data
model.load_and_clean_data('Absenteeism_new_data.csv')

# Check if predictions were calculated and appended to data correctly
model.predicted_outputs()

# Export data to csv for further analysis
model.predicted_outputs().to_csv('Absenteeism_predictions.csv', index = False)
